+++ app/main.py
@@ -4,8 +4,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from app.core.config import Settings
 from app.db.db import get_db
-from app.routers import health, user_routers
-from app.services import auth
+from app.routers import health, user_routers, auth_routers
 
 logging.basicConfig(
     filename='app.log',
@@ -43,7 +42,7 @@ async def shutdown_event():
 
 app.include_router(health.router)
 app.include_router(user_routers.user_router)
-app.include_router(auth.auth_router)
+app.include_router(auth_routers.auth_router)
 
 if __name__ == "__main__":
     uvicorn.run(
@@ -52,4 +51,3 @@ if __name__ == "__main__":
         port=Settings.ALLOWED_PORT,
         reload=True
     )
-
